requirements.txt
fastapi==0.115.8
fastapi[standard]
fastapi-users[beanie]
uvicorn==0.34.0
uvicorn[standard]
pydantic==2.10.6
pydantic-extra-types==2.10.2
passlib==1.7.4
passlib[bcrypt]
pyjwt==2.10.1
motor==3.7.0
motor[gssapi]
python-multipart==0.0.20
certifi==2025.1.31
pymongo==4.11.1
python-jose[cryptography]
beanie==1.29.0
beanie[srv]
python-dotenv==1.0.1
nh3==0.2.20